+++ src/use-async-task-delay.js
@@ -3,13 +3,11 @@ import shallowequal from 'shallowequal';
 
 import { useAsyncTaskTimeout } from './use-async-task-timeout';
 
-const emptyInputs = Symbol('emtpy inputs');
-
 export const useAsyncTaskDelay = (milliSeconds, inputs) => {
   const func = useRef();
-  const previousInputs = useRef(emptyInputs);
-  if (!shallowequal(inputs, previousInputs.current)) {
-    previousInputs.current = inputs;
+  const prevInputs = useRef();
+  if (!prevInputs.current || !shallowequal(prevInputs.current, inputs)) {
+    prevInputs.current = inputs;
     func.current = () => true;
   }
   return useAsyncTaskTimeout(func.current, milliSeconds);