+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [Unreleased]
+### Changed
+- Do not use useMemo/useCallback as a semantic guarantee
 
 ## [0.7.0] - 2019-01-22
 ### Added